% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_codes.R
\name{add_status_code_to_db}
\alias{add_status_code_to_db}
\title{Add Status Code to the Status Database}
\usage{
add_status_code_to_db(solver, code, symbol, message, roi_code = 1L)
}
\arguments{
\item{solver}{a character string giving the name of the solver.}

\item{code}{a integer giving the status code of the solver.}

\item{symbol}{a character string giving the status symbol.}

\item{message}{a character string used as status message.}

\item{roi_code}{a integer giving the ROI status code, 1L for failure and 0L for success.}
}
\description{
Add status a code to the status database.
}
\examples{
\dontrun{
add_status_code_to_db("ecos", 0L, "ECOS_OPTIMAL", "Optimal solution found.", 0L)
add_status_code_to_db(solver, -7L, "ECOS_FATAL", "Unknown problem in solver.", 1L)
add_status_code_to_db(solver, 5L, "GLP_OPT", "Solution is optimal.", 0L)
add_status_code_to_db(solver, 1L, "GLP_UNDEF", "Solution is undefined.", 1L)
}
}

