% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{as.constraint}
\alias{as.constraint}
\title{Constraint Utilities}
\usage{
as.constraint(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
an object inheriting from \code{"constraint"}.
}
\description{
Coerces objects of type \code{"constraint"}.
}
\author{
Stefan Theussl
}

