% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{as.F_objective}
\alias{as.F_objective}
\title{General Objective Function}
\usage{
as.F_objective(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
an object of class \code{"F_objective"} which inherits
from \code{"objective"}.
}
\description{
Coerces objects of type \code{"F_objective"}.
}
\details{
Objects from the following classes can be coerced to
\code{"F_objective"}: \code{"function"}, \code{"L_objective"}, and
\code{"Q_objective"}.
}
\author{
Stefan Theussl
}

