% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{as.L_constraint}
\alias{as.L_constraint}
\title{Linear Constraints}
\usage{
as.L_constraint(x, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{further arguments passed to or from other methods
(currently ignored).}
}
\value{
an object of class \code{"L_constraint"} which inherits
from \code{"constraint"}.
}
\description{
Coerces objects of type \code{"L_constraint"}.
}
\details{
Objects from the following classes can be coerced to
\code{"L_constraint"}: \code{"numeric"} and \code{"list"}. The
elements of a \code{"numeric"} vector \eqn{a} are treated as
objective variable coefficients of a single constraint in standard
form (\eqn{ax \geq 0}). A \code{"list"} must contain three
elements, the matrix \eqn{A}, the direction of constraints, and
the right hand side defining a linear constraint \eqn{Ax \leq b}.
}
\author{
Stefan Theussl
}

