% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{as.L_objective}
\alias{as.L_objective}
\title{Linear Objective Functions}
\usage{
as.L_objective(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
an object of class \code{"L_objective"} which inherits
from  \code{"Q_objective"} and \code{"objective"}.
}
\description{
Coerces objects of type \code{"L_objective"}.
}
\details{
Objects from the following classes can be coerced to
\code{"L_objective"}: \code{"NULL"}, \code{"numeric"},
\code{"Q_objective"}, and \code{"function"}. The elements of a
\code{"numeric"} vector \eqn{c} are treated as being objective
variable coefficients in \eqn{c^\top x}). Coercing from
\code{"Q_objective"} simply removes the quadratic part from the
objective function. Coercing a \code{"function"} to
\code{"L_objective"} is only possible if the function also
inherits from class \code{"objective"}.
}
\author{
Stefan Theussl
}

