% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{as.NO_constraint}
\alias{as.NO_constraint}
\title{Coercing Constraints}
\usage{
as.NO_constraint(x, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{further arguments passed to or from other methods
(currently ignored).}
}
\value{
an object of class \code{"NO_constraint"} which inherits
from \code{"constraint"}.
}
\description{
Coerces objects of type \code{"NO_constraint"}.
}
\details{
(Almost) all objects inheriting from class \code{"constraint"} can be
coerced to \code{"NO_constraint"}. It extracts the number of
(objective) variables from the original constraints objects and
returns the \code{"NO_constraint"} object.
}
\author{
Stefan Theussl
}

