% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{as.objective}
\alias{as.objective}
\title{Objective Function Utilities}
\usage{
as.objective(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
an object of class \code{"objective"}.
}
\description{
Coerces objects of type \code{"objective"}.
}
\author{
Stefan Theussl
}

