% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{as.Q_constraint}
\alias{as.Q_constraint}
\title{Quadratic Constraints}
\usage{
as.Q_constraint(x)
}
\arguments{
\item{x}{an R object.}

\item{...}{further arguments passed to or from other methods
(currently ignored).}
}
\value{
an object of class \code{"Q_constraint"} which inherits
from \code{"constraint"}.
}
\description{
Coerces objects of type \code{"Q_constraint"}.
}
\details{
Objects from the following classes can be coerced to
\code{"Q_constraint"}: \code{"list"}. The \code{"list"} must
contain four elements, a list of matrices \eqn{Q_m} representing
the quadratic part of \eqn{m} constraints, the matrix \eqn{A}
describing the linear part, the direction of the constraints, and
the right hand side.
}
\author{
Stefan Theussl
}

