% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{as.Q_objective}
\alias{as.Q_objective}
\title{Quadratic Objective Function}
\usage{
as.Q_objective(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
an object of class \code{"Q_objective"} which inherits
from \code{"objective"}.
}
\description{
Coerces objects of type \code{"Q_objective"}.
}
\details{
Objects from the following classes can be coerced to
\code{"Q_objective"}: \code{"function"}, \code{"matrix"}, and
\code{"simple_triplet_matrix"}.
}
\author{
Stefan Theussl
}

