\name{bounds.OP}
\alias{bounds.OP}
\title{Extract Objective Variable Bounds}
\usage{\method{bounds}{OP}(x)
}
\description{Extract bounds from ROI objects of class \code{"OP"} and return
them.}
\value{an object of class \code{"V_bound"}.}
\author{Stefan Theussl}
\arguments{\item{x}{an object of class \code{"OP"}.}
}

