\name{bounds.OP}
\alias{bounds.OP}
\title{Extract Objective Variable Bounds}
\usage{
  \method{bounds}{OP} (x)
}
\arguments{
  \item{x}{an object of class \code{"OP"}.}
}
\value{
  an object of class \code{"V_bound"}.
}
\description{
  Extract bounds from ROI objects of class \code{"OP"} and
  return them.
}
\author{
  Stefan Theussl
}

