% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{canonicalize_solution}
\alias{canonicalize_solution}
\title{Canonicalize Solution}
\usage{
canonicalize_solution(solution, optimum, status, solver, message = NULL, ...)
}
\arguments{
\item{solution}{a numeric or integer vector giving 
the solution of the optimization problem.}

\item{optimum}{a numeric giving the optimal value.}

\item{status}{an integer giving the status code (exit flag).}

\item{solver}{a character string giving the name of the solver.}

\item{message}{an optional \R object giving the original solver message.}

\item{...}{further arguments to be stored in the solution object.}
}
\value{
an object of class \code{"OP_solution"}.
}
\description{
Transform the solution to a standardized form.
}

