\name{constraints.OP}
\alias{constraints.OP}
\title{Extract constraints}
\usage{\method{constraints}{OP}(x)
}
\description{Extract constraints from ROI objects of class \code{"OP"} and
return them.}
\value{an object inheriting from class \code{"constraints"}.}
\author{Stefan Theussl}
\arguments{\item{x}{an object of class \code{"OP"}.}
}

