\name{constraints}
\alias{constraints}
\title{Extract constraints}
\usage{constraints(x)
}
\description{Extract constraints from its argument (typically ROI objects) and
return them.}
\details{Currently, there is no default method. See \code{\link{constraints.OP}}
for extracting constraints from ROI objects of class \code{"OP"}.}
\value{the extracted constraints object.}
\author{Stefan Theussl}
\arguments{\item{x}{an object used to select the method.}
}

