% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{F_objective}
\alias{F_objective}
\alias{as.F_objective}
\alias{terms.F_objective}
\alias{variable.names.F_objective}
\title{General (Nonlinear) Objective Function}
\usage{
F_objective(F, n, G = NULL, H = NULL, names = NULL)

\method{terms}{F_objective}(x, ...)

as.F_objective(x)

\method{variable.names}{F_objective}(object, ...)
}
\arguments{
\item{F}{an R \code{"function"} taking a numeric vector \code{x} of length \eqn{n} as argument.}

\item{n}{the number of objective variables.}

\item{G}{an R \code{"function"} returning the gradient at \code{x}.}

\item{H}{an optional \code{function} holding the Hessian of F.}

\item{names}{an optional character vector giving the names of x.}

\item{x}{an R object.}

\item{...}{further arguments passed to or from other methods}

\item{object}{an R object.}
}
\value{
an object of class \code{"F_objective"} which inherits
from \code{"objective"}.
}
\description{
General objective function \eqn{f(x)} to be optimized.
}
\author{
Stefan Theussl
}

