% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_solver_name}
\alias{get_solver_name}
\title{Get Solver Name}
\usage{
get_solver_name(pkgname)
}
\arguments{
\item{pkgname}{a string giving the package name.}
}
\value{
Returns the name of the solver as character.
}
\description{
Get the name of the solver plugin.
}

