\name{is.L_constraint}
\alias{is.L_constraint}
\title{Linear Constraints}
\usage{is.L_constraint(x)
}
\description{Tests if an object is interpretable as being of class \code{"L_constraint"}.}
\value{returns \code{TRUE} if its argument is of class
\code{"L_constraint"} and \code{FALSE} otherwise.}
\author{Stefan Theussl}
\arguments{\item{x}{object to be tested.}
}

