\name{is.Q_constraint}
\alias{is.Q_constraint}
\title{Quadratic Constraints}
\usage{is.Q_constraint(x)
}
\description{Tests if an object is interpretable as being of class \code{"Q_constraint"}.}
\value{returns \code{TRUE} if its argument is of class
\code{"Q_constraint"} and \code{FALSE} otherwise.}
\author{Stefan Theussl}
\arguments{\item{x}{object to be tested.}
}

