% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{is.default_bound}
\alias{is.default_bound}
\title{Check for default bounds}
\usage{
is.default_bound(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
a logical of length one indicating wether default bounds are given
}
\description{
tests if the given object is an variable bound
             which represents default values only 
             (i.e., all lower bounds are \code{0} 
             and all upper bounds as \code{Inf}).
}
