% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{is.F_constraint}
\alias{is.F_constraint}
\title{Function Constraints}
\usage{
is.F_constraint(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
returns \code{TRUE} if its argument is of class
\code{"F_constraint"} and \code{FALSE} otherwise.
}
\description{
Tests if an object is interpretable as being of class \code{"F_constraint"}.
}

