% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{is.NO_constraint}
\alias{is.NO_constraint}
\title{Class: \code{"NO_constraint"}}
\usage{
is.NO_constraint(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
returns \code{TRUE} if its argument is of class
\code{"NO_constraint"} and \code{FALSE} otherwise.
}
\description{
Tests if an object is interpretable as being of class \code{"L_constraint"}.
}
\author{
Stefan Theussl
}

