% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{J}
\alias{J}
\alias{J.L_constraint}
\alias{J.Q_constraint}
\title{Extract Jacobian Information}
\usage{
J(x, ...)

\method{J}{L_constraint}(x, ...)

\method{J}{Q_constraint}(x, ...)
}
\arguments{
\item{x}{a \code{\link{L_constraint}}, \code{\link{Q_constraint}} or 
\code{\link{F_constraint}}.}

\item{\ldots}{further arguments}
}
\value{
a list of functions
}
\description{
Derive the Jacobian for a given constraint.
}
\examples{
L <- matrix(c(3, 4, 2, 2, 1, 2, 1, 3, 2), nrow=3, byrow=TRUE)
lc <- L_constraint(L = L, dir = c("<=", "<=", "<="), rhs = c(60, 40, 80))
J(lc)
}
