\name{L_objective}
\alias{L_objective}
\title{Linear Objective Function}
\usage{L_objective(L)
}
\description{A linear objective function is typically of the form \eqn{c^\top
x} where \eqn{c} is a (sparse) vector of coefficients to the
\eqn{n} objective variables \eqn{x}.}
\author{Stefan Theussl}
\arguments{\item{L}{a numeric vector of length \eqn{n}, where \eqn{n} is the
number of objective variables.}
}

