% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{length.L_constraint}
\alias{length.L_constraint}
\title{Linear Constraints}
\usage{
\method{length}{L_constraint}(x)
}
\arguments{
\item{x}{constraints object.}
}
\value{
an integer.
}
\description{
Get the number of constraints from a corresponding ROI object.
}
\author{
Stefan Theussl
}

