% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{length.NO_constraint}
\alias{length.NO_constraint}
\title{Class: \code{"NO_constraint"}}
\usage{
\method{length}{NO_constraint}(x)
}
\arguments{
\item{x}{constraints object.}
}
\value{
an integer.
}
\description{
Get the number of constraints from a corresponding ROI object.
}
\author{
Stefan Theussl
}

