% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP.R
\name{OP_applicable_solver}
\alias{OP_applicable_solver}
\title{Applicable Solver}
\usage{
OP_applicable_solver(x)
}
\arguments{
\item{x}{an object of class \code{"OP"}}
}
\value{
A a character vector giving the giving the names of all available 
  and applicable solver
}
\description{
Takes an object of class \code{"OP"} (optimization problem) 
  and returns a character vector giving the names of all available 
  and applicable solver.
}

