% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP.R
\name{OP_signature}
\alias{OP_signature}
\title{Optimization Problem Signature}
\usage{
OP_signature(x)
}
\arguments{
\item{x}{an object of class \code{"OP"}}
}
\value{
A \code{data.frame} giving the signature of the
        the optimization problem.
}
\description{
Takes an object of class \code{"OP"} (optimization problem)
  and returns the signature of the optimization problem.
}
