% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{rbind.constraint}
\alias{rbind.constraint}
\title{Linear Constraints}
\usage{
rbind.constraint(..., use.names = FALSE, recursive = FALSE)
}
\arguments{
\item{...}{constraints objects to be concatenated.}

\item{use.names}{a logical if \code{FALSE} the names of the constraints
are ignored when combining them, if \code{TRUE} the constraints are
combined based on their \code{variable.names}.}

\item{recursive}{logical. Currently ignored (enable compatibility
with \code{c()} operator).}
}
\value{
an object of a class depending on the input which also
inherits from \code{"constraint"}. See \bold{Details}.
}
\description{
Take a sequence of constraints (ROI objects) arguments and combine
by rows, i.e., putting several constraints together.
}
\details{
The output type is determined from the highest type of the
components in the hierarchy \code{"L_constraint"} <
\code{"Q_constraint"} < \code{"F_constraint"}.
}
\author{
Stefan Theussl
}

