\name{ROI_available_solvers}
\alias{ROI_available_solvers}
\title{Solver Tools}
\usage{ROI_available_solvers(...)
}
\description{Retrieve all available solvers hosted in the given package repository.}
\value{a named character vector.}
\author{Stefan Theussl}
\arguments{\item{...}{arguments passed on to \code{\link{available.packages}}.}
}

