\name{ROI_installed_solvers}
\alias{ROI_installed_solvers}
\title{Solver Tools}
\usage{ROI_installed_solvers(...)
}
\description{Retrieve the names of all installed solvers.}
\value{a named character vector.}
\author{Stefan Theussl}
\arguments{\item{...}{arguments passed on to \code{\link{installed.packages}}.}
}

