% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROI_applicable_solvers}
\alias{ROI_applicable_solvers}
\title{Obtain Applicable Solvers}
\usage{
ROI_applicable_solvers(op)
}
\arguments{
\item{op}{an \pkg{ROI}-object of type \code{'OP'}.}
}
\value{
An character vector giving the applicable solver,
  for a certain optimization problem.
}
\description{
\code{ROI_applicable_solvers} takes as argument an
  optimization problem (object of class \code{'OP'}) and returns a vector
  giving the applicable solver. The set of applicable solver is restricted
  on the available solvers, which means if solver \code{"A"} and \code{"B"}
  would be applicable but a \code{ROI.plugin} is only installed for solver
  \code{"A"} only solver \code{"A"} would be listed as applicable solver.
}

