\name{ROI_available_solvers}
\alias{ROI_available_solvers}
\title{Solver Tools}
\usage{
  ROI_available_solvers(...)
}
\arguments{
  \item{...}{arguments passed on to
  \code{\link{available.packages}}.}
}
\value{
  a named character vector.
}
\description{
  Retrieve all available solvers hosted in the given
  package repository.
}
\author{
  Stefan Theussl
}

