% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bound (Constructors)}
\alias{bound (Constructors)}
\alias{c.bound}
\alias{is.bound}
\title{bound}
\usage{
\method{c}{bound}(...)

is.bound(x)
}
\arguments{
\item{x}{object to be tested}

\item{\ldots}{arguments (inheriting from bound) to be combined}
}
\description{
\pkg{ROI} distinguishes between 3 different 
  types of bounds: 
  \itemize{
  	\item No Bounds \code{NO_bound} 
  	\item Variable Bounds \code{\link{V_bound}} (inherits from \code{"bound"})
  	\item Conic Bounds \code{\link{C_bound}} (inherits from \code{"bound"})
  }
}
\details{
\pkg{ROI} provides the methods \code{\link{V_bound}} and 
  \code{\link{C_bound}} to be used as constructors for the corresponding bounds.
  \code{NO_bound} is not explicitly implemented but represented by \code{NULL}.
}

