% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{Bounds (Constructor)}
\alias{Bounds (Constructor)}
\title{Bounds}
\description{
\pkg{ROI} distinguishes between 3 different 
  types of bounds: 
  \itemize{
  	\item No Bounds \code{NO_bound} 
      (NO_bound is currently not explicitly implemented but 
       represented by \code{NULL})
  	\item Variable Bounds \code{\link{V_bound}}
  	\item Conic Bounds \code{\link{C_bound}}
  }
}
\details{
\pkg{ROI} offers provides the methods \code{NO_bound}, 
  \code{\link{C_bound}} and \code{\link{C_bound}} to be used as constructors
  for the corresponding bounds.
}

