% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint (Constructors)}
\alias{as.constraint}
\alias{c.constraint}
\alias{constraint (Constructors)}
\alias{dim.constraint}
\alias{is.constraint}
\title{constraint}
\usage{
\method{c}{constraint}(..., recursive = FALSE)

as.constraint(x)

is.constraint(x)

\method{dim}{constraint}(x)
}
\arguments{
\item{recursive}{a logical, giving if the arguments should be combined recursively.}

\item{x}{an object to be coerced or tested.}

\item{\ldots}{objects to be combined.}
}
\description{
\pkg{ROI} distinguishes between 4 different 
  types of constraint: 
  \itemize{
    \item No Constraint \code{\link{NO_constraint}} (inherits from \code{"constraint"})
    \item Linear Constraint \code{\link{L_constraint}} (inherits from \code{"constraint"})
    \item Quadratic Constraint \code{\link{Q_constraint}} (inherits from \code{"constraint"})
    \item Function Constraint \code{\link{F_constraint}} (inherits from \code{"constraint"})
  }
}

