\name{ROI_installed_solvers}
\alias{ROI_installed_solvers}
\title{Solver Tools}
\usage{
  ROI_installed_solvers(...)
}
\arguments{
  \item{...}{arguments passed on to
  \code{\link{installed.packages}}.}
}
\value{
  a named character vector.
}
\description{
  Retrieve the names of all installed solvers.
}
\author{
  Stefan Theussl
}

