% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROI_make_LP_signatures}
\alias{ROI_make_LP_signatures}
\alias{ROI_make_MILP_signatures}
\alias{ROI_make_MIQCP_signatures}
\alias{ROI_make_MIQP_signatures}
\alias{ROI_make_QP_signatures}
\title{Make Signatures}
\usage{
ROI_make_LP_signatures()

ROI_make_QP_signatures()

ROI_make_MILP_signatures()

ROI_make_MIQP_signatures()

ROI_make_MIQCP_signatures()
}
\value{
An R object containing the signature.
}
\description{
A set of default signatures.
}
\details{
TODO: Create one function with a parameter type or
  put them into the plugins!
}

