% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROI_make_signature}
\alias{ROI_make_signature}
\title{Make Signatures}
\usage{
ROI_make_signature(...)
}
\arguments{
\item{...}{signature definitions}
}
\value{
a data.frame with the supported signatures
}
\description{
Create a solver signature, the solver signatures are used to indicate
which problem types can be solved by a given solver.
}
\examples{
## ROI_make_LP_signatures
lp_signature <- ROI_make_signature( objective = "L",
                                    constraints = "L",
                                    types = c("C"),
                                    bounds = c("X", "V"),
                                    cones = c("free"),
                                    maximum = c(TRUE, FALSE) )
}

