% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ROI_options}
\alias{ROI_options}
\title{ROI Options}
\usage{
ROI_options(option, value)
}
\arguments{
\item{option}{any options can be defined, using 'key, value' pairs.
If 'value' is missing the current set value is returned for the given 'option'.
If both are missing. all set options are returned.}

\item{value}{the corresponding value to set for the given option.}
}
\description{
Allow the user to set and examine a variety of ROI options like the default
solver or the function used to compute the gradients.
}
