% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{ROI_plugin_get_solver_name}
\alias{ROI_plugin_get_solver_name}
\title{Get Solver Name}
\usage{
ROI_plugin_get_solver_name(pkgname)
}
\arguments{
\item{pkgname}{a string giving the package name.}
}
\value{
Returns the name of the solver as character.
}
\description{
Get the name of the solver plugin.
}
\seealso{
Other plugin functions: 
\code{\link{ROI_plugin_add_status_code_to_db}()},
\code{\link{ROI_plugin_build_equality_constraints}()},
\code{\link{ROI_plugin_build_inequality_constraints}()},
\code{\link{ROI_plugin_canonicalize_solution}()},
\code{\link{ROI_plugin_make_signature}()},
\code{\link{ROI_plugin_register_solver_control}()},
\code{\link{ROI_plugin_register_solver_method}()},
\code{\link{ROI_plugin_solution_prim}()},
\code{\link{ROI_registered_solver_control}()}
}
\concept{plugin functions}
