% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{.ROI_plugin_make_signature}
\alias{.ROI_plugin_make_signature}
\title{Make Signatures}
\usage{
.ROI_plugin_make_signature(...)
}
\arguments{
\item{...}{signature definitions}
}
\value{
a data.frame with the supported signatures
}
\description{
Create a solver signature, the solver signatures are used to indicate
which problem types can be solved by a given solver.
}
\examples{
## ROI_make_LP_signatures
lp_signature <- .ROI_plugin_make_signature( objective = "L",
                                    constraints = "L",
                                    types = c("C"),
                                    bounds = c("X", "V"),
                                    cones = c("free"),
                                    maximum = c(TRUE, FALSE) )
}
\seealso{
Other plugin.functions: \code{\link{.ROI_plugin_add_status_code_to_db}},
  \code{\link{.ROI_plugin_build_equality_constraints}},
  \code{\link{.ROI_plugin_build_inequality_constraints}},
  \code{\link{.ROI_plugin_canonicalize_solution}},
  \code{\link{.ROI_plugin_get_solver_name}},
  \code{\link{.ROI_plugin_register_solver_control}},
  \code{\link{.ROI_plugin_register_solver_method}}
}

