% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformulations.R
\name{ROI_plugin_register_reformulation}
\alias{ROI_plugin_register_reformulation}
\title{Register Reformulation Method}
\usage{
ROI_plugin_register_reformulation(from, to, method_name, method,
  description = "", cite = "", author = "")
}
\arguments{
\item{from}{a data.frame with the supported signatures.}

\item{to}{a data.frame with the supported signatures.}

\item{method_name}{a character string giving the name of the method.}

\item{method}{a function registered as solver method.}

\item{description}{a optional character string giving a description of what
the reformulation does.}

\item{cite}{a optional character string indicating a reference, 
such as the name of a book.}

\item{author}{a optional character string giving the name of the author.}
}
\value{
TRUE on success
}
\description{
Register a new reformulation method to be used with 
       \code{\link{ROI_reformulate}}.
}
\seealso{
Other reformulate functions: \code{\link{ROI_reformulate}},
  \code{\link{ROI_registered_reformulations}}
}

