% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{ROI_plugin_register_solver_control}
\alias{ROI_plugin_register_solver_control}
\title{Register Solver Controls}
\usage{
ROI_plugin_register_solver_control(solver, args, roi_control = "X")
}
\arguments{
\item{solver}{a character string giving the solver name.}

\item{args}{a character vector specifying with the supported signatures.}

\item{roi_control}{a character vector specifying the corresponding ROI control argument.}
}
\value{
TRUE on success
}
\description{
Register a new solver control argument.
}
\seealso{
Other plugin functions: 
\code{\link{ROI_plugin_add_status_code_to_db}()},
\code{\link{ROI_plugin_build_equality_constraints}()},
\code{\link{ROI_plugin_build_inequality_constraints}()},
\code{\link{ROI_plugin_canonicalize_solution}()},
\code{\link{ROI_plugin_get_solver_name}()},
\code{\link{ROI_plugin_make_signature}()},
\code{\link{ROI_plugin_register_solver_method}()},
\code{\link{ROI_plugin_solution_prim}()},
\code{\link{ROI_registered_solver_control}()}
}
\concept{plugin functions}
