% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{ROI_plugin_solution_prim}
\alias{ROI_plugin_solution_aux}
\alias{ROI_plugin_solution_dual}
\alias{ROI_plugin_solution_msg}
\alias{ROI_plugin_solution_prim}
\alias{ROI_plugin_solution_prim.OP_solution}
\alias{ROI_plugin_solution_prim.OP_solutions}
\alias{ROI_plugin_solution_psd}
\title{Extract solution from the solver.}
\usage{
ROI_plugin_solution_prim(x)

\method{ROI_plugin_solution_prim}{OP_solution}(x)

\method{ROI_plugin_solution_prim}{OP_solutions}(x)

ROI_plugin_solution_dual(x)

ROI_plugin_solution_aux(x)

ROI_plugin_solution_psd(x)

ROI_plugin_solution_msg(x)
}
\arguments{
\item{x}{an \code{R} object inheriting from \code{solution} or \code{solutions}.}
}
\value{
the corresponding solution/s.
}
\description{
Generic getter functions used by the function 
    \code{\link{solution}}. These functions can be used to write
    a solver specific getter function.
}
\seealso{
Other plugin functions: \code{\link{ROI_plugin_add_status_code_to_db}},
  \code{\link{ROI_plugin_build_equality_constraints}},
  \code{\link{ROI_plugin_build_inequality_constraints}},
  \code{\link{ROI_plugin_canonicalize_solution}},
  \code{\link{ROI_plugin_get_solver_name}},
  \code{\link{ROI_plugin_make_signature}},
  \code{\link{ROI_plugin_register_solver_control}},
  \code{\link{ROI_plugin_register_solver_method}}
}

