% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tools.R
\name{ROI_read}
\alias{ROI_read}
\title{Read Optimization Problems}
\usage{
ROI_read(file, type, solver = NULL, ...)
}
\arguments{
\item{file}{a character giving the name of the file the optimization problem
is to be read from.}

\item{type}{a character giving the type of the file 
(e.g. \code{"mps_free"}, \code{"mps_fixed"}, \code{"lp_cplex"},
      \code{"lp_lpsolve"}, ...).}

\item{solver}{an optional character giving the name of the plugin 
(e.g. \code{"lpsolve"}).}

\item{...}{further arguments passed on to the read method.}
}
\value{
x an optimization problem of class \code{"OP"}.
}
\description{
Reads an optimization problem from various file formats and
             returns an optimization problem of class \code{"OP"}.
}
\seealso{
Other input output: \code{\link{ROI_plugin_register_reader_writer}},
  \code{\link{ROI_registered_reader}},
  \code{\link{ROI_registered_writer}},
  \code{\link{ROI_write}}
}
\concept{input output}
