% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformulations.R
\name{ROI_reformulate}
\alias{ROI_reformulate}
\title{Reformulate a Optimization Problem}
\usage{
ROI_reformulate(x, to, method = NULL)
}
\arguments{
\item{x}{an object of type \code{'OP'} giving the optimization problem.}

\item{to}{a data.frame with the supported signatures.}

\item{method}{a character string giving the name of the method.}
}
\value{
the reformulated optimization problem.
}
\description{
Register a new reformulation method.
}
\seealso{
Other reformulate functions: \code{\link{ROI_plugin_register_reformulation}}
}

