% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROI_register_solver_method}
\alias{ROI_register_solver_method}
\title{Register Solver Method}
\usage{
ROI_register_solver_method(signatures, solver, method)
}
\arguments{
\item{signatures}{a data.frame with the supported signatures.}

\item{solver}{a character string giving the solver name.}

\item{method}{a function registered as solver method.}
}
\value{
TRUE on success
}
\description{
Register a new solver method.
}

