% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tools.R
\name{ROI_registered_reader}
\alias{ROI_registered_reader}
\title{List Registered Reader}
\usage{
ROI_registered_reader(type = NULL)
}
\arguments{
\item{type}{an optional character giving the type of the file 
(e.g. \code{"mps_free"}, \code{"mps_fixed"}, \code{"lp_cplex"},
      \code{"lp_lpsolve"}, ...).}
}
\value{
x a data.frame containing information about the registered readers.
}
\description{
Retrieve meta information about the registered reader
}
\examples{
ROI_registered_reader()
ROI_registered_reader("mps_fixed")
}
\seealso{
Other input output: \code{\link{ROI_plugin_register_reader_writer}},
  \code{\link{ROI_registered_writer}},
  \code{\link{read.op}}, \code{\link{write.op}}
}
