% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformulations.R
\name{ROI_registered_reformulations}
\alias{ROI_registered_reformulations}
\title{Registered Reformulations}
\usage{
ROI_registered_reformulations()
}
\value{
a data.frame giving some information about the registered reformulation 
    methods.
}
\description{
Retrieve meta information about the registered reformulations.
}
\examples{
ROI_registered_reformulations()
}
\seealso{
Other reformulate functions: \code{\link{ROI_plugin_register_reformulation}},
  \code{\link{ROI_reformulate}}
}
\concept{reformulate functions}
