% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{ROI_registered_solver_control}
\alias{ROI_registered_solver_control}
\title{Registered Solver Controls}
\usage{
ROI_registered_solver_control(solver)
}
\arguments{
\item{solver}{a character string giving the solver name.}
}
\value{
a \code{data.frame} giving the control arguments.
}
\description{
Retrieve the registered solver control arguments.
}
\seealso{
Other plugin functions: 
\code{\link{ROI_plugin_add_status_code_to_db}()},
\code{\link{ROI_plugin_build_equality_constraints}()},
\code{\link{ROI_plugin_build_inequality_constraints}()},
\code{\link{ROI_plugin_canonicalize_solution}()},
\code{\link{ROI_plugin_get_solver_name}()},
\code{\link{ROI_plugin_make_signature}()},
\code{\link{ROI_plugin_register_solver_control}()},
\code{\link{ROI_plugin_register_solver_method}()},
\code{\link{ROI_plugin_solution_prim}()}
}
\concept{plugin functions}
