% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tools.R
\name{ROI_registered_writer}
\alias{ROI_registered_writer}
\title{Write Optimization Problems}
\usage{
ROI_registered_writer(signature = NULL)
}
\arguments{
\item{signature}{an optimization problem of class \code{"OP"}.}
}
\description{
Write an optimization problem to file.
}
\examples{
ROI_registered_writer()
op <- OP(1:2)
ROI_registered_writer(OP_signature(op))
}
\seealso{
Other input output: \code{\link{ROI_plugin_register_reader_writer}},
  \code{\link{ROI_registered_reader}},
  \code{\link{read.op}}, \code{\link{write.op}}
}
\concept{input output}
