% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROI_require_solver}
\alias{ROI_require_solver}
\title{Require Solver}
\usage{
ROI_require_solver(solver, warn = 0)
}
\arguments{
\item{solver}{a character string giving the solver name.}

\item{warn}{an integer giving if the warn level.
For \code{warn = -1} the warning is ignored. For \code{warn = 0}
the warning is stored and printed later. For \code{warn = 1} the
warning is printed immediately. For \code{warn = 2} the
warning is turned into an error. Default is \code{warn = 0}.}
}
\value{
Returns \code{TRUE} on success otherwise \code{FALSE}.
}
\description{
Loads the specified solver and registers it in an internal data base.
A request to load an already loaded solver has no effect.
}
