% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{ROI_solver_signature}
\alias{ROI_solver_signature}
\title{Obtain Solver Signature}
\usage{
ROI_solver_signature(solver)
}
\arguments{
\item{solver}{a character string giving the name of the solver.}
}
\value{
the solver signature if the specified solver is registered \code{NULL} otherwise.
}
\description{
Obtain the signature of a registered solver.
}
\examples{
ROI_solver_signature("nlminb")
}

