% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_tools.R
\name{ROI_write}
\alias{ROI_write}
\title{Write Optimization Problems}
\usage{
ROI_write(x, file, type, solver = NULL, ...)
}
\arguments{
\item{x}{an optimization problem of class \code{"OP"}.}

\item{file}{a character giving the name of the file the optimization problem
is to be written.}

\item{type}{a character giving the type of the file 
(e.g. \code{"freemps"}, \code{"mps_fixed"}, \code{"lp_cplex"},
      \code{"lp_lpsolve"}, ...).}

\item{solver}{an optional character giving the name of the plugin 
(e.g. \code{"lpsolve"}).}

\item{...}{further arguments passed on to the write method.}
}
\description{
Write an optimization problem to file.
}
\seealso{
Other input output: 
\code{\link{ROI_plugin_register_reader_writer}},
\code{\link{ROI_read}()},
\code{\link{ROI_registered_reader}()},
\code{\link{ROI_registered_writer}()}
}
\concept{input output}
