% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{constraint directions}
\alias{constraint directions}
\alias{eq}
\alias{leq}
\alias{geq}
\title{Replicate \code{"=="}, \code{">="} and \code{"<="} Signs}
\usage{
eq(n)

leq(n)

geq(n)
}
\arguments{
\item{n}{an integer giving the number of times the sign should be repeated.}
}
\description{
The utility functions \code{eq}, \code{leq} and 
  \code{geq} replicate the signs \code{"=="}, \code{">="} and \code{"<="}
  \code{n} times.
}
\examples{
eq(3)
leq(2)
geq(4)
}
