% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vech}
\alias{vech}
\title{Half Vectorization}
\usage{
vech(...)
}
\arguments{
\item{...}{one or more matrices to be half-vectorized.}
}
\value{
a matrix
}
\description{
The utility function \code{vech} performs a
  half-vectorization on the given matrices.
}

