% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{C_bound}
\alias{C_bound}
\alias{as.C_bound}
\alias{is.C_bound}
\title{Conic Bounds}
\usage{
C_bound(..., type = c("free", "nonneg", "soc", "psd", "expp", "expd", "powp",
  "powd"))

as.C_bound(x)

is.C_bound(x)
}
\arguments{
\item{...}{arguments which give the row index in the A matrix.}

\item{type}{an character giving the type of the bound, 
valid types are \code{"free"}, \code{"nonneg"}, \code{"soc"}, \code{"psd"}, 
\code{"expp"}, \code{"expd"}, \code{"powp"} and \code{"powd"}.}

\item{x}{object to be coerced or tested.}
}
\value{
An S3 object of class \code{"C_bound"} containing the conic
        bounds.
}
\description{
Construct a conic bounds object.
}
\examples{
conic_bounds <- c(C_bound(1, type="free"), C_bound(2:4, type="soc"))
cones <- list("free"=c(1), "soc"=list(2:4))
conic_bounds <- as.C_bound(cones)
cones <- list("free"=c(1, 2), "expd"=list(3:5))
bound <- as.C_bound(cones)
}

