% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{terms.C_constraint}
\alias{C_constraint}
\alias{as.C_constraint}
\alias{is.C_constraint}
\alias{length.C_constraint}
\alias{terms.C_constraint}
\alias{variable.names.C_constraint}
\title{Linear Constraints}
\usage{
\method{terms}{C_constraint}(x, ...)

C_constraint(L, cones, rhs, names = NULL)

as.C_constraint(x, ...)

is.C_constraint(x)

\method{length}{C_constraint}(x)

\method{variable.names}{C_constraint}(object, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{further arguments passed to or from other methods
(currently ignored).}

\item{L}{a numeric vector of length \eqn{n} (a single constraint)
or a matrix of dimension \eqn{n \times m}, where \eqn{n} is the
number of objective variables and \eqn{m} is the number of
constraints. Matrices can be of class
\code{"simple_triplet_matrix"} to allow a sparse representation of
constraints.}

\item{cones}{TODO}

\item{rhs}{a numeric vector with the right hand side of the constraints.}

\item{names}{an optional character vector giving the names of \eqn{x} 
(column names of \eqn{A}).}

\item{object}{an R object.}
}
\value{
an object of class \code{"L_constraint"} which inherits
        from \code{"constraint"}.
}
\description{
Linear constraints are typically of the form \deqn{Lx \leq rhs}
where \eqn{L} is a \eqn{m \times n} (sparse) matrix of coefficients 
to the objective variables \eqn{x} and the right hand side \eqn{rhs} 
is a vector of length \eqn{m}.
}

