% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\title{Extract constraints}
\usage{
constraints(x)
}
\arguments{
\item{x}{an object used to select the method.}
}
\value{
the extracted constraints object.
}
\description{
Extract constraints from its argument (typically ROI objects) and
return them.
}
\details{
Currently, there is no default method. See \code{\link{constraints.OP}}
for extracting constraints from ROI objects of class \code{"OP"}.
}
\author{
Stefan Theussl
}

