% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints (Set/Get)}
\alias{constraints (Set/Get)}
\alias{constraints}
\alias{constraints.OP}
\alias{constraints<-}
\title{Constraints - Accessor and Mutator Functions}
\usage{
constraints(x)

\method{constraints}{OP}(x)

constraints(x) <- value
}
\arguments{
\item{x}{an object used to select the method.}

\item{value}{an R object.}
}
\value{
the extracted constraints object.
}
\description{
The \link{constraints} of a given optimization problem (\link{OP}) 
    can be accessed or mutated via the method \code{'constraints'}.
}
\examples{
## minimize: x + 2 y
## subject to: x + y >= 1
## x, y >= 0
x <- OP(1:2)
constraints(x) <- L_constraint(c(1, 1), ">=", 1)
constraints(x)
}
\author{
Stefan Theussl
}
