% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{F_objective}
\alias{F_objective}
\title{General (Nonlinear) Objective Function}
\usage{
F_objective(F, n, G = NULL)
}
\arguments{
\item{F}{an R \code{"function"} taking a numeric vector \code{x} of length \eqn{n} as argument.}

\item{n}{the number of objective variables.}

\item{G}{an R \code{"function"} returning the gradient at \code{x}.}
}
\value{
an object of class \code{"F_objective"} which inherits
from \code{"objective"}.
}
\description{
General objective function \eqn{f(x)}to be optimized.
}
\author{
Stefan Theussl
}

