% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{is.L_constraint}
\alias{is.L_constraint}
\title{Linear Constraints}
\usage{
is.L_constraint(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
returns \code{TRUE} if its argument is of class
\code{"L_constraint"} and \code{FALSE} otherwise.
}
\description{
Tests if an object is interpretable as being of class \code{"L_constraint"}.
}
\author{
Stefan Theussl
}

