% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum.R
\name{maximum}
\alias{maximum}
\alias{maximum<-}
\title{Extract Direction}
\usage{
maximum(x)

maximum(x) <- value
}
\arguments{
\item{x}{an object used to select the method.}

\item{value}{an R object.}
}
\value{
a logical giving the direction.
}
\description{
Extract the direction of the optimization problem, i.e.
             if we want to minimize or maximize the objective.
}
\author{
Stefan Theussl
}

