% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{NO_constraint}
\alias{NO_constraint}
\title{Class: \code{"NO_constraint"}}
\usage{
NO_constraint(n_obj)
}
\arguments{
\item{n_obj}{a numeric vector of length \code{1} representing the number
of objective variables.}
}
\value{
an object of class \code{"NO_constraint"} which inherits
from \code{"L_constraint"} and \code{"constraint"}.
}
\description{
In case the \code{constraints} slot in the problem object is
\code{NULL} the return value of a call of \code{constraints()}
will return an object of class \code{"NO_constraint"} which
inherits from \code{"L_constraint"}.
}
\author{
Stefan Theussl
}

