% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective}
\alias{as.objective}
\alias{objective}
\alias{objective<-}
\title{Extract Objective Functions}
\usage{
objective(x)

objective(x) <- value

as.objective(x)
}
\arguments{
\item{x}{an object used to select the method.}

\item{value}{an R object.}
}
\value{
a function inheriting from \code{"objective"}.
}
\description{
Extract the objective function from its argument (typically ROI
objects) and return them.
}
\details{
The default method assumes that the supplied R object is a list
where the element \code{objective} represents the objective
function. The extractet element is then coerced to a function.
}
\author{
Stefan Theussl
}

