% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{objective (Set/Get)}
\alias{objective (Set/Get)}
\alias{objective}
\alias{objective<-}
\alias{as.objective}
\title{Objective - Accessor and Mutator Functions}
\usage{
objective(x)

objective(x) <- value

as.objective(x)
}
\arguments{
\item{x}{an object used to select the method.}

\item{value}{an R object.}
}
\value{
a function inheriting from \code{"objective"}.
}
\description{
The \link{objective} of a given optimization problem (\link{OP}) 
    can be accessed or mutated via the method \code{'objective'}.
}
\examples{
x <- OP()
objective(x) <- 1:3
}
\author{
Stefan Theussl
}
