% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROI_registered_solvers}
\alias{ROI_available_solvers}
\alias{ROI_installed_solvers}
\alias{ROI_registered_solvers}
\title{Solver Tools}
\usage{
ROI_registered_solvers(...)

ROI_installed_solvers(...)

ROI_available_solvers(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link{installed.packages}}.}
}
\value{
a named character vector.
}
\description{
Retrieve the names of installed or registered solvers.
}
\details{
Whereas \code{ROI_installed_solvers()} and
  \code{ROI_available_solvers()} may list the names of installed
  solvers that do not necessarily work,
  \code{ROI_registered_solvers()} lists all solvers that can be used
  to solve optimization problems.
}
\author{
Stefan Theussl
}

