% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{ROI_solve}
\alias{ROI_solve}
\title{Solve an Optimization Problem}
\usage{
ROI_solve(x, solver = "", control = list(), ...)
}
\arguments{
\item{x}{an optimization problem of class \code{"OP"}.}

\item{solver}{a character vector specifying the solver to use.}

\item{control}{a list with additional control parameters for the solver.
This is solver specific so please consult the corresponding documentation.}

\item{...}{a list of control parameters (overruling those specified in \code{control}).}
}
\value{
a list containing the solution and a message from the solver.
}
\description{
Solve a given optimization problem.
  This function uses the given solver (or searches for an appropriate
  solver) to solve the supplied optimization problem.
}
\author{
Stefan Theussl
}

