% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{solution}
\alias{solution}
\title{Extract Solution}
\usage{
solution(x, type = c("primal", "dual", "aux", "psd", "msg"))
}
\arguments{
\item{x}{an object of type \code{'OP_solution'}.}

\item{type}{sdfdf}
}
\value{
the extracted solution.
}
\description{
The solution can be accessed via the method \code{'solution'}.
}

