% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{solution}
\alias{solution}
\title{Extract Solution}
\usage{
solution(x, type = c("primal", "dual", "aux", "psd", "msg"), ...)
}
\arguments{
\item{x}{an object of type \code{'OP_solution'}.}

\item{type}{a character giving the name of the solution to be extracted.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the extracted solution.
}
\description{
The solution can be accessed via the method \code{'solution'}.
}

