% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{types (Set/Get)}
\alias{types}
\alias{types (Set/Get)}
\alias{types<-}
\title{Types - Accessor and Mutator Functions}
\usage{
types(x)

types(x) <- value
}
\arguments{
\item{x}{an object used to select the method.}

\item{value}{an R object.}
}
\value{
a character vector.
}
\description{
The \link{types} of a given optimization problem (\link{OP}) 
    can be accessed or mutated via the method \code{'types'}.
}
\examples{
## minimize: x + 2 y
## subject to: x + y >= 1
## x, y >= 0    x, y are integer
x <- OP(objective = 1:2, constraints = L_constraint(c(1, 1), ">=", 1))
types(x) <- c("I", "I")
types(x)
}
\author{
Stefan Theussl
}

